moved locations to be conformant with standard
