+++ tests/test_issues/input/issue_1104_data.yaml
@@ -14,3 +14,4 @@ concrete_thing_set:
     description: the sevenhundredeightyninth named thing
     alternative_identifiers:
       - nmdc:sevenhundredeightyninth
+
