+++ tests/test_issues/input/issue_1104_schema.yaml
