+++ tests/test_issues/model/issue_1104_classes.py
